Only installs fonts if they haven't already been installed

+++ vimconfig.sh
@@ -7,9 +7,15 @@ git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 cp configs/.vimrc ~/.vimrc
 # Download patched fonts for airline (special status bar at the bottom)
 # Note: you will have to change the font of your terminal manually to actually make use of the fonts. The fonts as a whole are called "Powerline patched", so the font that I might load would be "Ubuntu Mono derivative Powerline Regular." I find that that font at size 13 is a suitable replacement in gnome-terminal, the default Ubuntu terminal. More documentation can be found in the README.md for the powerline fonts. 
-git clone https://github.com/powerline/fonts.git
-./fonts/install.sh
-rm -rf fonts
+
+#If the patched fonts for vim-airline have not been installed (as evidenced by
+#the presence of the Ubuntu font), then install them.
+if [ -f $HOME/.local/share/fonts/'Ubuntu Mono derivative Powerline.ttf']; then
+	git clone https://github.com/powerline/fonts.git
+	./fonts/install.sh
+	rm -rf fonts
+fi
+
 # Use Vundle to install all the programs called for by .vimrc
 vi +PluginInstall +qall
 # Makes vim the default manpager (mostly for aesthetic purposes, but vim is such a functional home too)
